use seperated ActivityContextManager class & pre-pair testing stuff for fixing group member number and name location
