+++ PlayerTags/Features/TagTargetFeature.cs
@@ -15,50 +15,25 @@ namespace PlayerTags.Features
     /// </summary>
     public abstract class TagTargetFeature : IDisposable
     {
-        private ActivityContext m_CurrentActivityContext;
+        public ActivityContextManager ActivityContextManager { get; init; }
 
         public TagTargetFeature()
         {
-
-            m_CurrentActivityContext = ActivityContext.None;
-
-            PluginServices.ClientState.TerritoryChanged += ClientState_TerritoryChanged;
+            ActivityContextManager = new();
         }
 
         public virtual void Dispose()
         {
-            PluginServices.ClientState.TerritoryChanged -= ClientState_TerritoryChanged;
+            ActivityContextManager.Dispose();
         }
 
         protected abstract bool IsIconVisible(Tag tag);
 
         protected abstract bool IsTextVisible(Tag tag);
 
-        private void ClientState_TerritoryChanged(object? sender, ushort e)
-        {
-            m_CurrentActivityContext = ActivityContext.None;
-
-            var contentFinderConditionsSheet = PluginServices.DataManager.GameData.GetExcelSheet<ContentFinderCondition>();
-            if (contentFinderConditionsSheet != null)
-            {
-                var foundContentFinderCondition = contentFinderConditionsSheet.FirstOrDefault(contentFinderCondition => contentFinderCondition.TerritoryType.Row == PluginServices.ClientState.TerritoryType);
-                if (foundContentFinderCondition != null)
-                {
-                    if (foundContentFinderCondition.PvP)
-                    {
-                        m_CurrentActivityContext = ActivityContext.PvpDuty;
-                    }
-                    else
-                    {
-                        m_CurrentActivityContext = ActivityContext.PveDuty;
-                    }
-                }
-            }
-        }
-
         protected bool IsTagVisible(Tag tag, GameObject? gameObject)
         {
-            bool isVisibleForActivity = ActivityContextHelper.GetIsVisible(m_CurrentActivityContext,
+            bool isVisibleForActivity = ActivityContextHelper.GetIsVisible(ActivityContextManager.CurrentActivityContext,
                 tag.IsVisibleInPveDuties.InheritedValue ?? false,
                 tag.IsVisibleInPvpDuties.InheritedValue ?? false,
                 tag.IsVisibleInOverworld.InheritedValue ?? false);
